renamed utiltities.js file to reportUtility as it only contained reporting related functions
